remove free-standing fn proc macro, replace from_env by get, use a sealed type to run tests
