+++ crates/astria-composer/Cargo.toml
@@ -8,7 +8,7 @@ rust-version = "1.70.0"
 
 [dependencies]
 proto = { package = "astria-proto", path = "../astria-proto" }
-astria-config = { path = "../astria-config", features = ["config", "derive"] }
+astria-config = { path = "../astria-config", features = ["derive"] }
 telemetry = { package = "astria-telemetry", path = "../astria-telemetry" }
 
 ethers = { version = "2.0.7", features = ["ws"] }
