+++ crates/astria-conductor/Cargo.toml
@@ -31,7 +31,7 @@ tokio-util = { workspace = true, features = ["rt"] }
 tonic = { workspace = true }
 tracing = { workspace = true }
 
-astria-config = { path = "../astria-config", features = ["config", "derive"] }
+astria-config = { path = "../astria-config", features = ["derive"] }
 sequencer-client = { package = "astria-sequencer-client", path = "../astria-sequencer-client", features = [
   "http",
   "websocket",
