+++ crates/astria-sequencer-relayer/Cargo.toml
@@ -33,7 +33,7 @@ tendermint-rpc = { workspace = true, features = ["http-client"] }
 tracing = { workspace = true }
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 
-astria-config = { path = "../astria-config", features = ["config", "derive"] }
+astria-config = { path = "../astria-config", features = ["derive"] }
 astria-celestia-jsonrpc-client = { path = "../astria-celestia-jsonrpc-client" }
 astria-sequencer-validation = { path = "../astria-sequencer-validation" }
 proto = { package = "astria-proto", path = "../astria-proto" }
