+++ crates/astria-sequencer-relayer/src/main.rs
@@ -1,13 +1,14 @@
+use astria_config::AstriaConfig as _;
 use astria_sequencer_relayer::{
-    config::get_config,
     telemetry,
+    Config,
     SequencerRelayer,
 };
 use tracing::info;
 
 #[tokio::main]
 async fn main() {
-    let cfg = get_config().expect("failed to read configuration");
+    let cfg = Config::get().expect("failed to read configuration");
     telemetry::init(std::io::stdout, &cfg.log).expect("failed to setup telemetry");
     info!(
         config = serde_json::to_string(&cfg).expect("serializing to a string cannot fail"),
